// src/generated-api/models/Attack.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Underworld
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.6.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    AttackEffect,
    AttackEffectFromJSON,
    AttackEffectFromJSONTyped,
    AttackEffectToJSON,
} from './AttackEffect';

/**
 * 
 * @export
 * @interface Attack
 */
export interface Attack {
    /**
     * 
     * @type {number}
     * @memberof Attack
     */
    num_rolls: number;
    /**
     * 
     * @type {number}
     * @memberof Attack
     */
    modifier: number;
    /**
     * 
     * @type {Array<AttackEffect>}
     * @memberof Attack
     */
    effects: Array<AttackEffect>;
}

export function AttackFromJSON(json: any): Attack {
    return AttackFromJSONTyped(json, false);
}

export function AttackFromJSONTyped(json: any, ignoreDiscriminator: boolean): Attack {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'num_rolls': json['num_rolls'],
        'modifier': json['modifier'],
        'effects': ((json['effects'] as Array<any>).map(AttackEffectFromJSON)),
    };
}

export function AttackToJSON(value?: Attack | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'num_rolls': value.num_rolls,
        'modifier': value.modifier,
        'effects': ((value.effects as Array<any>).map(AttackEffectToJSON)),
    };
}