// src/generated-api/models/Consumable.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Underworld
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.6.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    ConsumableEffect,
    ConsumableEffectFromJSON,
    ConsumableEffectFromJSONTyped,
    ConsumableEffectToJSON,
} from './ConsumableEffect';

/**
 * 
 * @export
 * @interface Consumable
 */
export interface Consumable {
    /**
     * 
     * @type {number}
     * @memberof Consumable
     */
    uses: number;
    /**
     * 
     * @type {boolean}
     * @memberof Consumable
     */
    knows_uses: boolean;
    /**
     * 
     * @type {ConsumableEffect}
     * @memberof Consumable
     */
    effect: ConsumableEffect;
    /**
     * 
     * @type {boolean}
     * @memberof Consumable
     */
    knows_effect: boolean;
}

export function ConsumableFromJSON(json: any): Consumable {
    return ConsumableFromJSONTyped(json, false);
}

export function ConsumableFromJSONTyped(json: any, ignoreDiscriminator: boolean): Consumable {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'uses': json['uses'],
        'knows_uses': json['knows_uses'],
        'effect': ConsumableEffectFromJSON(json['effect']),
        'knows_effect': json['knows_effect'],
    };
}

export function ConsumableToJSON(value?: Consumable | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'uses': value.uses,
        'knows_uses': value.knows_uses,
        'effect': ConsumableEffectToJSON(value.effect),
        'knows_effect': value.knows_effect,
    };
}