// src/generated-api/models/Defense.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Underworld
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.6.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Defense
 */
export interface Defense {
    /**
     * 
     * @type {number}
     * @memberof Defense
     */
    damage_resistance: number;
}

export function DefenseFromJSON(json: any): Defense {
    return DefenseFromJSONTyped(json, false);
}

export function DefenseFromJSONTyped(json: any, ignoreDiscriminator: boolean): Defense {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'damage_resistance': json['damage_resistance'],
    };
}

export function DefenseToJSON(value?: Defense | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'damage_resistance': value.damage_resistance,
    };
}