// src/generated-api/models/MovePlayerItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Underworld
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EquipLocationTag,
    EquipLocationTagFromJSON,
    EquipLocationTagFromJSONTyped,
    EquipLocationTagToJSON,
} from './EquipLocationTag';

/**
 * 
 * @export
 * @interface MovePlayerItem
 */
export interface MovePlayerItem {
    /**
     * 
     * @type {string}
     * @memberof MovePlayerItem
     */
    item_id: string;
    /**
     * 
     * @type {EquipLocationTag}
     * @memberof MovePlayerItem
     */
    location_tag: EquipLocationTag;
    /**
     * 
     * @type {boolean}
     * @memberof MovePlayerItem
     */
    put_at_the_ready: boolean;
}

export function MovePlayerItemFromJSON(json: any): MovePlayerItem {
    return MovePlayerItemFromJSONTyped(json, false);
}

export function MovePlayerItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): MovePlayerItem {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'item_id': json['item_id'],
        'location_tag': EquipLocationTagFromJSON(json['location_tag']),
        'put_at_the_ready': json['put_at_the_ready'],
    };
}

export function MovePlayerItemToJSON(value?: MovePlayerItem | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'item_id': value.item_id,
        'location_tag': EquipLocationTagToJSON(value.location_tag),
        'put_at_the_ready': value.put_at_the_ready,
    };
}